requirements.txt
streamlit
langchain
ollama
langchain-ollama
streamlit-audiorecorder
speechrecognition
gtts